requirements.txt
loguru==0.6.0
prometheus-client==0.15.0
semver==2.13.0
art==5.8
pyyaml==6.0
python-gitlab==3.12.0
